+++ part6/redux-anecdotes/src/App.js
@@ -1,38 +1,12 @@
-import { useSelector, useDispatch } from 'react-redux'
-import { addVote } from './reducers/anecdoteReducer'
 import AnecdoteForm from './components/AnecdoteForm'
+import AnecdoteList from './components/AnecdoteList'
 
 const App = () => {
-  const anecdotes = useSelector(state => state.sort(function(a, b) {
-    if (a.votes > b.votes) {
-      return -1
-    } else if (a.votes < b.votes){
-      return 1
-    } else {
-      return 0
-    }
-  }))
-  const dispatch = useDispatch()
-
   
-  const vote = (id) => {
-    dispatch(addVote(id))
-  }
-
   return (
     <div>
       <h2>Anecdotes</h2>
-      {anecdotes.map(anecdote =>
-        <div key={anecdote.id}>
-          <div>
-            {anecdote.content}
-          </div>
-          <div>
-            has {anecdote.votes}
-            <button onClick={() => vote(anecdote.id)}>vote</button>
-          </div>
-        </div>
-      )}
+      <AnecdoteList />
       <AnecdoteForm />
     </div>
   )
