+++ part6/redux-anecdotes/src/components/Anecdote.js
@@ -1,0 +1,15 @@
+const Anecdote = ({ anecdote, handleClick }) => {
+  return (
+    <div>
+      <div>
+        {anecdote.content}
+      </div>
+      <div>
+        has {anecdote.votes}
+        <button onClick={handleClick}>vote</button>
+      </div>
+    </div>
+  )
+}
+
+export default Anecdote
